[test] Get rid of utils.py in aot python scripts
Moving functions into each file so that it's easier to read.
